requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.4.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0